+++ tests/unit/experimental/components/test_kg_writer.py
@@ -26,10 +26,8 @@ from neo4j_graphrag.experimental.components.types import (
     Neo4jRelationship,
 )
 from neo4j_graphrag.neo4j_queries import (
-    UPSERT_NODE_QUERY,
-    UPSERT_NODE_QUERY_VARIABLE_SCOPE_CLAUSE,
-    UPSERT_RELATIONSHIP_QUERY,
-    UPSERT_RELATIONSHIP_QUERY_VARIABLE_SCOPE_CLAUSE,
+    upsert_node_query,
+    upsert_relationship_query,
 )
 
 
@@ -56,11 +54,16 @@ def test_batched() -> None:
     return_value=None,
 )
 def test_upsert_nodes(_: Mock, driver: MagicMock) -> None:
+    driver.execute_query.return_value = (
+        [{"element_id": "#1"}],
+        None,
+        None,
+    )
     neo4j_writer = Neo4jWriter(driver=driver)
     node = Neo4jNode(id="1", label="Label", properties={"key": "value"})
     neo4j_writer._upsert_nodes(nodes=[node], lexical_graph_config=LexicalGraphConfig())
     driver.execute_query.assert_called_once_with(
-        UPSERT_NODE_QUERY,
+        upsert_node_query(False),
         parameters_={
             "rows": [
                 {
@@ -88,6 +91,11 @@ def test_upsert_nodes_with_embedding(
     _: Mock,
     driver: MagicMock,
 ) -> None:
+    driver.execute_query.return_value = (
+        [{"element_id": "#1"}],
+        None,
+        None,
+    )
     neo4j_writer = Neo4jWriter(driver=driver)
     node = Neo4jNode(
         id="1",
@@ -95,10 +103,9 @@ def test_upsert_nodes_with_embedding(
         properties={"key": "value"},
         embedding_properties={"embeddingProp": [1.0, 2.0, 3.0]},
     )
-    driver.execute_query.return_value.records = [{"elementId(n)": 1}]
     neo4j_writer._upsert_nodes(nodes=[node], lexical_graph_config=LexicalGraphConfig())
     driver.execute_query.assert_any_call(
-        UPSERT_NODE_QUERY,
+        upsert_node_query(False),
         parameters_={
             "rows": [
                 {
@@ -130,7 +137,9 @@ def test_upsert_relationship(_: Mock, driver: MagicMock) -> None:
         type="RELATIONSHIP",
         properties={"key": "value"},
     )
-    neo4j_writer._upsert_relationships(rels=[rel])
+    neo4j_writer._upsert_relationships(
+        rels=[rel],
+    )
     parameters = {
         "rows": [
             {
@@ -143,7 +152,7 @@ def test_upsert_relationship(_: Mock, driver: MagicMock) -> None:
         ]
     }
     driver.execute_query.assert_called_once_with(
-        UPSERT_RELATIONSHIP_QUERY,
+        upsert_relationship_query(False),
         parameters_=parameters,
         database_=None,
     )
@@ -167,7 +176,9 @@ def test_upsert_relationship_with_embedding(_: Mock, driver: MagicMock) -> None:
         embedding_properties={"embeddingProp": [1.0, 2.0, 3.0]},
     )
     driver.execute_query.return_value.records = [{"elementId(r)": "rel_elem_id"}]
-    neo4j_writer._upsert_relationships(rels=[rel])
+    neo4j_writer._upsert_relationships(
+        rels=[rel],
+    )
     parameters = {
         "rows": [
             {
@@ -180,7 +191,7 @@ def test_upsert_relationship_with_embedding(_: Mock, driver: MagicMock) -> None:
         ]
     }
     driver.execute_query.assert_any_call(
-        UPSERT_RELATIONSHIP_QUERY,
+        upsert_relationship_query(False),
         parameters_=parameters,
         database_=None,
     )
@@ -196,13 +207,21 @@ def test_upsert_relationship_with_embedding(_: Mock, driver: MagicMock) -> None:
     return_value=None,
 )
 async def test_run(_: Mock, driver: MagicMock) -> None:
+    driver.execute_query.return_value = (
+        [
+            {"element_id": "#1"},
+            {"element_id": "#2"},
+        ],
+        None,
+        None,
+    )
     neo4j_writer = Neo4jWriter(driver=driver)
     node = Neo4jNode(id="1", label="Label")
     rel = Neo4jRelationship(start_node_id="1", end_node_id="2", type="RELATIONSHIP")
     graph = Neo4jGraph(nodes=[node], relationships=[rel])
     await neo4j_writer.run(graph=graph)
     driver.execute_query.assert_any_call(
-        UPSERT_NODE_QUERY,
+        upsert_node_query(False),
         parameters_={
             "rows": [
                 {
@@ -228,7 +247,7 @@ async def test_run(_: Mock, driver: MagicMock) -> None:
         ]
     }
     driver.execute_query.assert_any_call(
-        UPSERT_RELATIONSHIP_QUERY,
+        upsert_relationship_query(False),
         parameters_=parameters_,
         database_=None,
     )
@@ -242,7 +261,14 @@ async def test_run(_: Mock, driver: MagicMock) -> None:
 async def test_run_is_version_below_5_23(_: Mock) -> None:
     driver = MagicMock()
     driver.execute_query = Mock(
-        return_value=([{"versions": ["5.22.0"], "edition": "enterprise"}], None, None)
+        side_effect=(
+            # get_version
+            ([{"versions": ["5.22.0"], "edition": "enterpise"}], None, None),
+            # upsert nodes
+            ([{"_internal_id": "1", "element_id": "#1"}], None, None),
+            # upsert relationships
+            (None, None, None),
+        )
     )
 
     neo4j_writer = Neo4jWriter(driver=driver)
@@ -253,7 +279,7 @@ async def test_run_is_version_below_5_23(_: Mock) -> None:
     await neo4j_writer.run(graph=graph)
 
     driver.execute_query.assert_any_call(
-        UPSERT_NODE_QUERY,
+        upsert_node_query(False),
         parameters_={
             "rows": [
                 {
@@ -279,7 +305,7 @@ async def test_run_is_version_below_5_23(_: Mock) -> None:
         ]
     }
     driver.execute_query.assert_any_call(
-        UPSERT_RELATIONSHIP_QUERY,
+        upsert_relationship_query(False),
         parameters_=parameters_,
         database_=None,
     )
@@ -293,7 +319,14 @@ async def test_run_is_version_below_5_23(_: Mock) -> None:
 async def test_run_is_version_5_23_or_above(_: Mock) -> None:
     driver = MagicMock()
     driver.execute_query = Mock(
-        return_value=([{"versions": ["5.23.0"], "edition": "enterpise"}], None, None)
+        side_effect=(
+            # get_version
+            ([{"versions": ["5.23.0"], "edition": "enterpise"}], None, None),
+            # upsert nodes
+            ([{"element_id": "#1"}], None, None),
+            # upsert relationships
+            (None, None, None),
+        )
     )
 
     neo4j_writer = Neo4jWriter(driver=driver)
@@ -305,7 +338,7 @@ async def test_run_is_version_5_23_or_above(_: Mock) -> None:
     await neo4j_writer.run(graph=graph)
 
     driver.execute_query.assert_any_call(
-        UPSERT_NODE_QUERY_VARIABLE_SCOPE_CLAUSE,
+        upsert_node_query(True),
         parameters_={
             "rows": [
                 {
@@ -331,7 +364,7 @@ async def test_run_is_version_5_23_or_above(_: Mock) -> None:
         ]
     }
     driver.execute_query.assert_any_call(
-        UPSERT_RELATIONSHIP_QUERY_VARIABLE_SCOPE_CLAUSE,
+        upsert_relationship_query(True),
         parameters_=parameters_,
         database_=None,
     )